fix : 메세지 페이지 폴더 이동 및 개인 테스트 폴더 삭제
