--- frontend/src/TestGang.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const TestGang = () => {
-  return <></>;
-};
-
-export default TestGang;
