+++ frontend/src/pages/Home/Home.jsx
@@ -1,10 +1,5 @@
 import React from "react";
-import TestSung from "../../TestSung";
 
 export default function Home() {
-  return (
-    <div>
-      <TestSung />
-    </div>
-  );
+  return <div>home</div>;
 }
