+++ frontend/src/pages/Project/Project.jsx
@@ -1,10 +1,5 @@
 import React from "react";
-import TestGang from "../../TestGang";
 
 export default function Project() {
-  return (
-    <div>
-      <TestGang />
-    </div>
-  );
+  return <div>projects</div>;
 }
